Fix init script

+++ init.sh
@@ -6,7 +6,19 @@ if [ $# -ne 1 ]; then
 fi
 
 dirname=$1
-mkdir -v $dirname
-cp -v Makefile $dirname/
-cp -v sample.c $dirname/
-ln -sf ../run.sh $dirname/run.sh
+
+if [ ! -e $dirname ]; then
+  mkdir -v $dirname
+fi
+
+if [ ! -e $dirname/Makefile ]; then
+  cp -v Makefile $dirname/
+fi
+
+if [ ! -e $dirname/sample.c ]; then
+  cp -v sample.c $dirname/
+fi
+
+if [ ! -L $dirname/sample.c ]; then
+  ln -sf ../run.sh $dirname/run.sh
+fi
